taskApi/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
dj-database-url==3.1.2
Django==5.2.1
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==26.0.0
kombu==5.6.2
packaging==26.2
pillow==12.2.0
prompt_toolkit==3.0.52
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.7.0
psycopg2-binary==2.9.10f